[TS-302] Refactor: habitRecordConstants 삭제 homeConstants로 통합
